# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "FCC API",
        "description": "This is a online API Documentation for FCC",
        "contact": {
            "name": "CJ"
        },
        "version": "1.0.0"
    },
    "host": "192.168.1.14:8000",
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/app/settings": {
            "get": {
                "tags": [
                    "Guests"
                ],
                "summary": "Get settings(Rules & General)",
                "description": "",
                "operationId": "api.guest.settings",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Server Error"
                    },
                    "400": {
                        "description": "Logic Error"
                    }
                }
            }
        },
...
    },
...
}